renaming folder
папка на гите с маленькой буквы,
